+++ .github/workflows/lint-build.yml
@@ -27,5 +27,5 @@ jobs:
       - run: yarn install
       - name: Lint Code
         run: yarn lint
-      - name: Build frontend
-        run: yarn build:frontend
+      - name: Build
+        run: yarn build
