+++ backend/config/admin.js
@@ -5,4 +5,9 @@ module.exports = ({ env }) => ({
   apiToken: {
     salt: env("API_TOKEN_SALT"),
   },
+  transfer: {
+    token: {
+      salt: env("TRANSFER_TOKEN_SALT"),
+    },
+  },
 });
