+++ backend/package.json
@@ -13,18 +13,18 @@
   },
   "dependencies": {
     "@_sh/strapi-plugin-ckeditor": "^1.1.1",
-    "@strapi-community/strapi-provider-upload-google-cloud-storage": "^4.3.6",
-    "@strapi/plugin-graphql": "4.3.1",
-    "@strapi/plugin-i18n": "4.3.1",
-    "@strapi/plugin-users-permissions": "4.3.1",
-    "@strapi/strapi": "4.3.1",
+    "@strapi-community/strapi-provider-upload-google-cloud-storage": "^4.6.1",
+    "@strapi/plugin-graphql": "4.9.1",
+    "@strapi/plugin-i18n": "4.9.1",
+    "@strapi/plugin-users-permissions": "4.9.1",
+    "@strapi/strapi": "4.9.1",
     "better-sqlite3": "7.6.2",
-    "knex": "^2.4.0",
-    "pg": "^8.7.3",
+    "knex": "^2.4.2",
+    "pg": "^8.9.0",
     "pg-connection-string": "^2.5.0",
-    "sqlite3": "^5.1.2",
-    "strapi-connector-bookshelf": "^3.6.10",
-    "strapi-plugin-import-export-entries": "^1.7.1"
+    "sqlite3": "^5.1.6",
+    "strapi-connector-bookshelf": "^3.6.11",
+    "strapi-plugin-import-export-entries": "^1.18.1"
   },
   "author": {
     "name": "A Strapi developer"
