+++ backend/public/.gitignore
@@ -1,1 +1,2 @@
 uploads/*
+!uploads/.gitkeep
