+++ config/development.env
@@ -6,6 +6,7 @@ APP_KEYS=testkey1,testkey2
 DATABASE_CLIENT=sqlite
 DATABASE_FILENAME=sami-dev.db
 API_TOKEN_SALT=test_salt
+TRANSFER_TOKEN_SALT=test_salt
 ADMIN_JWT_SECRET=test_secret
 JWT_SECRET=dGVzdF9qd3Rfc2VjcmV0
