+++ config/staging.env
@@ -6,6 +6,7 @@ APP_KEYS=testkey1,testkey2
 API_TOKEN_SALT=test_salt
 ADMIN_JWT_SECRET=test_secret
 JWT_SECRET=dGVzdF9qd3Rfc2VjcmV0
+TRANSFER_TOKEN_SALT=test_salt
 
 # Database 
 DATABASE_CLIENT=postgres
