+++ package.json
@@ -9,7 +9,7 @@
     "shared"
   ],
   "scripts": {
-    "build": "yarn scripts dev build",
+    "build": "yarn scripts cli build",
     "deploy:frontend": "vercel --cwd frontend",
     "export:frontend": "yarn workspace frontend export",
     "format": "yarn format:code && yarn format:style",
@@ -20,9 +20,8 @@
     "lint:code": "eslint . --ext .js,.jsx,.ts,.tsx --color",
     "lint:spell:TODO": "npx cspell \"**/*.md\" --config ./.cspell.json",
     "lint:style": "prettier --check '**/*.{md,json,js,tsx,ts}'",
-    "release": "release-it",
     "scripts": "yarn workspace scripts start",
-    "start": "yarn scripts dev start"
+    "start": "yarn scripts cli start"
   },
   "devDependencies": {
     "@release-it/conventional-changelog": "^5.0.0",
@@ -39,9 +38,8 @@
     "lint-staged": "^12.4.1",
     "prettier": "^2.7.1",
     "pretty-quick": "^3.1.3",
-    "release-it": "^15.0.0",
-    "typescript": "4.8.4",
-    "vercel": "^28.18.3"
+    "vercel": "^28.18.3",
+    "typescript": "4.8.4"
   },
   "husky": {
     "hooks": {
