+++ scripts/package.json
@@ -3,11 +3,11 @@
   "private": true,
   "main": "src/commands/index.ts",
   "scripts": {
-    "start": "ts-node src/commands/index.ts",
+    "start": "ts-node-esm src/commands/index.ts",
     "dev": "ts-node-dev --respawn --watch src/**/*.ts src/commands/index.ts"
   },
   "dependencies": {
-    "@google-cloud/storage": "^6.9.4",
+    "@google-cloud/storage": "^6.9.5",
     "axios": "^0.27.2",
     "backend": "workspace:*",
     "boxen": "^5.0.1",
@@ -15,21 +15,23 @@
     "chokidar": "^3.5.3",
     "commander": "^8.3.0",
     "concurrently": "^7.5.0",
-    "dotenv": "^16.0.1",
-    "fs-extra": "^9.0.1",
-    "glob": "^7.2.0",
+    "dotenv": "^16.0.3",
+    "execa": "^5.1.1",
+    "fs-extra": "^11.1.1",
+    "glob": "^10.2.1",
     "inquirer": "^7.3.3",
     "knex-schema-inspector": "^2.0.4",
+    "pm2": "^5.3.0",
     "prompts": "^2.4.2",
     "typescript": "^4.1.2",
     "wait-on": "^6.0.1"
   },
   "devDependencies": {
-    "@types/fs-extra": "^9.0.4",
-    "@types/inquirer": "^7.3.1",
-    "@types/node": "^18.6.3",
-    "@types/prompts": "^2.0.14",
-    "ts-node": "~7.0.0",
-    "ts-node-dev": "^1.0.0"
+    "@types/fs-extra": "^11.0.1",
+    "@types/inquirer": "^7.3.3",
+    "@types/node": "^18.15.11",
+    "@types/prompts": "^2.4.4",
+    "ts-node": "~10.9.1",
+    "ts-node-dev": "^2.0.0"
   }
 }
