+++ scripts/src/commands/cli/build.ts
