+++ scripts/src/commands/cli/setEnv.ts
