+++ scripts/src/commands/cli/start.ts
@@ -1,10 +1,9 @@
 import { Command } from "commander";
 import type { ConcurrentlyCommandInput } from "concurrently";
 import concurrently from "concurrently";
-import { existsSync } from "fs";
 import { resolve } from "path";
 import { PATHS } from "../../paths";
-import { logError, loadEnv } from "../../utils";
+import { loadEnv } from "../../utils";
 import type { IEnvLoaded } from "../../utils";
 
 /***************************************************************************************
@@ -40,31 +39,12 @@ class StartCmd {
     // when running frontend always assume local config
     const frontendStart = this.getFrontendCommand();
     const { result } = concurrently([backendStart, frontendStart], {
-      killOthers: ["failure", "success"],
+      ["killOthers" as any]: ["failure", "success"],
     });
     await result;
   }
 
   private getBackendStartCommand(envLoaded: IEnvLoaded): ConcurrentlyCommandInput {
-    const { STRAPI_READONLY_TOKEN, GOOGLE_APPLICATION_CREDENTIALS } = envLoaded.parsed;
-    // ensure frontend bootstrapped
-
-    if (!STRAPI_READONLY_TOKEN) {
-      logError({ msg1: "Strapi must be bootstrapped first, run command:", msg2: `yarn scripts strapi bootstrap` });
-    }
-    // ensure external storage configured
-    if (GOOGLE_APPLICATION_CREDENTIALS) {
-      const serviceAccountPath = resolve(PATHS.configDir, GOOGLE_APPLICATION_CREDENTIALS);
-      if (!existsSync(serviceAccountPath)) {
-        logError({
-          msg1: "Google application credentials not found",
-          msg2: serviceAccountPath,
-        });
-      }
-      // rewrite as absolute path
-      process.env.GOOGLE_APPLICATION_CREDENTIALS = serviceAccountPath;
-    }
-
     console.log("starting backend...");
     return {
       name: "strapi",
